+++ src/role/role.js
@@ -1,3 +1,12 @@
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
 import { Card } from "../card/card.js";
 import { Rank } from "../card/rank.enum.js";
 import { Suit } from "../card/suit.enum.js";
@@ -48,12 +57,12 @@ export class Role {
         return (this.hand.length === 2) && (Rules.BLACK_JACK === this.score);
     }
     clearHand() {
-        var _a;
-        this.hand = [];
-        this.writeScore();
-        document.querySelectorAll(`.${(_a = this.role) === null || _a === void 0 ? void 0 : _a.id}-card`).forEach(card => card.className = 'clear-hand');
-        setTimeout(() => {
-            this.role.innerHTML = '';
-        }, 1e3);
+        return __awaiter(this, void 0, void 0, function* () {
+            var _a;
+            this.hand = [];
+            this.writeScore();
+            document.querySelectorAll(`.${(_a = this.role) === null || _a === void 0 ? void 0 : _a.id}-card`).forEach(card => card.className = 'clear-hand');
+            return new Promise(resolve => setTimeout(resolve, 1e3)).then(() => this.role.innerHTML = '');
+        });
     }
 }